+++ bpaf_derive/src/lib.rs
@@ -3,6 +3,8 @@ use quote::ToTokens;
 use syn::parse_macro_input;
 
 mod field;
+#[cfg(test)]
+mod field_tests;
 mod kw;
 mod top;
 mod utils;
